requirements.txt
Flask==2.3.3
gunicorn==20.1.0
opencv-python==4.10.0.84
numpy==1.24.3
requests
python-socketio[client]
eventlet